requirements.txt
numpy>=1.20.0
sentence-transformers>=2.2.0
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
